+++ src/main/java/chess/domain/piece/QuadrantPiece.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Move;
+import chess.domain.move.Move;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
