+++ src/test/java/chess/domain/AbstractTestFixture.java
@@ -1,5 +1,10 @@
 package chess.domain;
 
+import chess.domain.move.Direction;
+import chess.domain.move.Move;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import java.util.List;
 
 public abstract class AbstractTestFixture {
