+++ src/test/java/chess/domain/move/DirectionTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.move;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -13,14 +13,14 @@ class DirectionTest {
     @CsvSource({"UP,DOWN", "DOWN,UP", "LEFT,LEFT", "RIGHT,RIGHT"})
     void flipHorizontal(Direction oldDirection, Direction newDirection) {
         assertThat(oldDirection.flipHorizontal())
-            .isEqualTo(newDirection);
+                .isEqualTo(newDirection);
     }
-    
+
     @DisplayName("좌우 대칭 할 수 있다.")
     @ParameterizedTest(name = "{0}을 좌우 대칭하면 {1}이다.")
     @CsvSource({"UP,UP", "DOWN,DOWN", "LEFT,RIGHT", "RIGHT,LEFT"})
     void flipVertical(Direction oldDirection, Direction newDirection) {
         assertThat(oldDirection.flipVertical())
-            .isEqualTo(newDirection);
+                .isEqualTo(newDirection);
     }
 }
