+++ src/test/java/chess/domain/move/MoveTest.java
@@ -1,12 +1,14 @@
-package chess.domain;
+package chess.domain.move;
 
-import static chess.domain.Direction.DOWN;
-import static chess.domain.Direction.LEFT;
-import static chess.domain.Direction.RIGHT;
-import static chess.domain.Direction.UP;
+import static chess.domain.move.Direction.DOWN;
+import static chess.domain.move.Direction.LEFT;
+import static chess.domain.move.Direction.RIGHT;
+import static chess.domain.move.Direction.UP;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.domain.AbstractTestFixture;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
